Comentando e estrutrando código

+++ include/NoVertice.h
@@ -19,17 +19,17 @@ public:
 
     // Getters e Setters
     int getIdVertice();                                     // Retorna o identificador do vertice
-    int getPesoVertice();                                   // Retorna o peso do vertice
+    float getPesoVertice();                                 // Retorna o peso do vertice
     int getNumArestas();                                    // Retorna a quantidade de arestas do vertice
     NoVertice* getProximo();                                // Retorna o proximo no vertice
     ListaAdjAresta* getArestas();                           // Retorna a lista encadeada de arestas
     
     void setProximo(NoVertice* proximo);                    // Define o proximo no vertice
-    int setIdVertice(int novoId);                                     // Define o identificador do vertice
+    int setIdVertice(int novoId);                           // Define o identificador do vertice
     
     // Funcoes auxiliares
     int getNumVizinhos();                                   // Retorna o numero de vizinhos do vertice (numero de arestas de saida/ grau do vertice)
-    void adicionar_aresta(int id, float peso);                // Adiciona uma aresta ao vertice
+    void adicionar_aresta(int id, float peso);              // Adiciona uma aresta ao vertice
     void remover_aresta(int destino);                       // Remove uma aresta do vertice
     void remover_primeira_aresta();                         // Remove a primeira aresta do vertice
 };
