+++ src/NoAresta.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-// Construtor
+// #region Construtor e Destrutor
 NoAresta::NoAresta(int idVerticeOrigem, int idVerticeDestino, float peso) {
     this->idVerticeOrigem = idVerticeOrigem;
     this->idVerticeDestino = idVerticeDestino;
@@ -11,35 +11,42 @@ NoAresta::NoAresta(int idVerticeOrigem, int idVerticeDestino, float peso) {
     this->proximo = nullptr;
 }
 
-// Destrutor
 NoAresta::~NoAresta() {
 
 }
+// #endregion
 
+// Retorna o vertice de origem da aresta
 int NoAresta::getOrigem() {
     return idVerticeOrigem;
 }
 
+// Retorna o vertice de destino da aresta
 int NoAresta::getDestino() {
     return idVerticeDestino;
 }
 
-int NoAresta::getPeso() {
+// Retorna o peso da aresta
+float NoAresta::getPeso() {
     return peso;
 }
 
+// Retorna o proximo Noh aresta
 NoAresta* NoAresta::getProximo() {
     return proximo;
 }
 
+// Define o proximo Noh aresta
 void NoAresta::setProximo(NoAresta* proximo) {
     this->proximo = proximo;
 }
 
+// Define o id do vertice de origem
 void NoAresta::setVerticeOrigem(int novoId) {
     this->idVerticeOrigem = novoId;
 }
 
+// Define o id do vertice de destino
 void NoAresta::setVerticeDestino(int novoId) {
     this->idVerticeDestino = novoId;
 }
